SSCSCI-481: revert localDevServer changes to restore logs to appInsights
