+++ charts/sscs-tribunals-frontend/Chart.yaml
@@ -3,7 +3,7 @@ name: sscs-tribunals-frontend
 home: https://github.com/hmcts/sscs-submit-your-appeal
 apiVersion: v2
 appVersion: "1.0"
-version: 0.2.40
+version: 0.2.41
 maintainers:
   - name: HMCTS SSCS Team
 dependencies:
