+++ charts/sscs-tribunals-frontend/values.yaml
@@ -9,21 +9,16 @@ nodejs:
     UPLOAD_EVIDENCE_URL: http://sscs-tribunals-api-{{ .Values.global.environment }}.service.core-compute-{{ .Values.global.environment }}.internal/evidence/upload
     SERVICES_IDAM_LOGIN_URL: https://idam-web-public.{{ .Values.global.environment }}.platform.hmcts.net/login
     SERVICES_IDAM_API_URL: https://idam-api.{{ .Values.global.environment }}.platform.hmcts.net
-
     # seems to be disabled currently and only enabled locally pointing at AAT
     # SERVICES_WEB_FORM:
-
     # This is currently disabled in prod, enabled all other environments
     CONTACT_US_WEB_FORM_ENABLED: false
     MEDIA_FILES_ALLOWED_ENABLED: true
-
     POSTCODE_CHECKER_ALLOWED_RPCS: "birmingham,liverpool,sutton,leeds,newcastle,cardiff,glasgow,bradford"
     UV_THREADPOOL_SIZE: 64
     CONTACT_US_WEBCHAT_ENABLED: false
     FT_WELSH: true
     ALLOW_DLA_ENABLED: true
-    ALLOW_CONFIG_MUTATIONS: true
-
     #Added for new Webchat with Agent
     WEBCHAT_URL: "webchat.ctsc.hmcts.net"
     WEBCHAT_CLIENT_URL: "webchat-client.ctsc.hmcts.net"
