+++ package.json
@@ -38,8 +38,6 @@
     "lint"
   ],
   "dependencies": {
-    "@azure/identity": "^3.3.0",
-    "@azure/keyvault-secrets": "^4.7.0",
     "@babel/core": "^7.22.11",
     "@babel/preset-env": "^7.22.10",
     "@cypress/request": "^3.0.0",
